ec: bump host_event response struct to 64-bit
Needed to send 64-bit host event masks to AP.

Update ectool accordingly to print 64-bit masks.

BUG=b:69329196
TEST=None

+++ include/ec_commands.h
@@ -3018,7 +3018,7 @@ struct __ec_align4 ec_params_host_event_mask {
 };
 
 struct __ec_align4 ec_response_host_event_mask {
-	uint32_t mask;
+	uint64_t mask;
 };
 
 /* These all use ec_response_host_event_mask */
